chore(test): refactor release-test code
no functional change, just splitting it into multiple files
as it had grown into a beast, plus renaming a couple of suites
to clarify the main difference in terms of VPCs / subnets
